Solved even more bugs. One of them made the cursor in low range chart be wrong (outside screen) when coming from galactic chart and a far planet was selected.
Added FLG_PIRATE to Thargoids, so police helps.

Tweaked circle drawing so no clipping is done if not necessary. Can improve this more by inlining things, but as we are low on memory I prefer to wait before doing this.
Removed several clcs and secs which should not be needed to speed up drawing.

Also changed a couple of things in memory layout in obj3d.s and LineDraw.s, to gain a good amount of over 200 bytes. 

Removed USE_ACCURATE_CLIPPING from the params.h file, as it speeds things up, reduces code and does not seem to change anything.

Added a #define USE_OBLETS to remove all the code related to Compound object drawing, to save memory and speed up the code (just a bit).

Moved the object records area (over 2K) to fffa-size.