Added the possibility of plotting circles in short range chart (with a define, so they are not active by default), just for testing. Should work better with filled circles.
For hyperspace circles: added a check to avoid a ship entering hyper if another ship is entering hyper (in tactics.s)

Cleaned a bit tineloop.s and added a check in case the circles when hyperspacing are too big.

Seems it is working quite well now.

Small tab adjusting in universe.s